# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PT Multi Daya Mitra Company Profile CMS API
  version: 1.2.0
  summary: REST API for the public corporate website and protected CMS.
  description: |
    Sign-in is two-step: POST /auth/login verifies the password and returns
    either tokens (2FA disabled or a valid trusted-device token was presented)
    or an OTP challenge completed via POST /auth/verify-otp. Analytics
    collection is first-party and feature-flagged via the "analytics" setting.
servers:
  - url: http://localhost:8080/api/v1
    description: Local Docker API
security: []
...
paths:
...
  /admin/media:
    get:
      tags: [Media]
      security:
        - bearerAuth: []
      summary: List uploaded media files, newest first.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: q
          in: query
          schema:
            type: string
          description: Filter by file name.
      responses:
        "200":
          description: Paginated media.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaList"
...
  /admin/media/{id}:
    delete:
      tags: [Media]
      security:
        - bearerAuth: []
      summary: Permanently delete a media file (row + stored object).
      parameters:
        - $ref: "#/components/parameters/UUID"
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/redirects:
    get:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: List short links (q search; status active|inactive|expired).
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: q
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [active, inactive, expired] }
      responses:
        "200":
          description: Paginated redirects with lifetime scan totals.
    post:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Create a short link.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedirectInput"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redirect"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /admin/redirects/{id}:
    get:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Get one short link.
      parameters:
        - $ref: "#/components/parameters/UUID"
      responses:
        "200":
          description: Redirect.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Redirect"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Update a short link (optimistic locking).
      parameters:
        - $ref: "#/components/parameters/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedirectInput"
      responses:
        "200":
          description: Updated.
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Redirects]
      security:
        - bearerAuth: []
      summary: Archive a short link (stops redirecting; scan history kept).
      parameters:
        - $ref: "#/components/parameters/UUID"
        - name: version
          in: query
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Archived.
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UUID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: perPage
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        perPage:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    MediaFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        objectKey:
          type: string
        url:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        createdAt:
          type: string
          format: date-time
    MediaList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MediaFile"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    Redirect:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        destination: { type: string }
        description: { type: string }
        redirectType: { type: integer, enum: [301, 302] }
        isActive: { type: boolean }
        expiresAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        version: { type: integer }
        totalScans: { type: integer, description: Lifetime scans from hourly rollups. }
    RedirectInput:
      type: object
      required: [name, slug, destination, redirectType]
      properties:
        name: { type: string }
        slug:
          type: string
          pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
          description: Must not collide with site routes, system pages, or CMS page keys.
        destination: { type: string, format: uri }
        description: { type: string }
        redirectType: { type: integer, enum: [301, 302] }
        isActive: { type: boolean }
        expiresAt: { type: [string, "null"], format: date-time }
        version: { type: integer, description: Required on update. }
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Slug conflict or optimistic locking conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...